% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_decay_fn.R
\name{add_decay_fn}
\alias{add_decay_fn}
\title{Add A Decaying Rescue Effect to Post-Rescue Therapy Measurements}
\usage{
add_decay_fn(y, phi, d, decay)
}
\arguments{
\item{y}{A numeric vector of longitudinal outcome measurements.}

\item{phi}{Numeric. A parameter that controls the rate of decay.}

\item{d}{A numeric vector that indicates the distance between each
longitudinal measurement and the time of rescue therapy.}

\item{decay}{Character. What decay function to use? Options are
\code{"linear"}, \code{"exponential"}, or \code{"gaussian"}.}
}
\description{
Add A Decaying Rescue Effect to Post-Rescue Therapy Measurements
}
