% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_design_matrix.R
\name{build_design_matrix}
\alias{build_design_matrix}
\title{Design Matrices for Longitudinal RCT}
\usage{
build_design_matrix(
  N_t,
  N_c,
  K,
  time_start = 0,
  time_scale = 1,
  interaction = TRUE,
  num_binary_covariates = 0,
  success_probs = 0.5,
  num_continuous_covariates = 0,
  return_type = "matrix"
)
}
\arguments{
\item{N_t}{Numeric. Number of patients that receive active treatment.}

\item{N_c}{Numeric. Number of patients that receive control treatment.}

\item{K}{Numeric. The number of longitudinal measurements.}

\item{time_start}{Numeric. Defines the starting time of follow-up.}

\item{time_scale}{Numeric. A value multiplied by the time vector to scale times of measurement.}

\item{interaction}{Logical. When \code{TRUE} includes an interaction term between treatment and time.}

\item{num_binary_covariates}{Numeric. The number of binary covariates to draw from Bernoulli distribution.}

\item{success_probs}{Numeric vector of success probabilities for binary covariates.}

\item{num_continuous_covariates}{Numeric. The number of continuous covariates to draw from a standard
Normal distribution.}

\item{return_type}{Character. Incicates whether to return a \code{"matrix"} or \code{"data.frame"}.}
}
\value{
A matrix or data frame.
}
\description{
Builds a design matrix for a 2-group RCT assuming longitudinal follow-up.
}
\note{
This function produces a design matrix that can be used to generate longitudinal outcome data. That is,
it is a first step in simulating longitudinal RCT data. If no covariates are used then it may be
produced a single time and does not need to be recreated anew for each data set. If covariates are included
the intention of the simulation comes into play, i.e., generating data anew for each data set implies that
the simulations sample new individuals, whereas generating once assumes we generate new outcomes for the
same individuals.
}
\examples{
# fixed effects design for treated
xmat_trt <- build_design_matrix(
  N_t = 1,
  N_c = 0,
  K = 5,
  time_start = 0,
  time_scale = 1/4
)

# fixed effects design for control
xmat_ctrl <- build_design_matrix(
  N_t = 0,
  N_c = 1,
  K = 5,
  time_start = 0,
  time_scale = 1/4
)


}
