% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_lmm_cov.R
\name{build_lmm_cov}
\alias{build_lmm_cov}
\title{Build Linear Mixed Model Covariance Matrix}
\usage{
build_lmm_cov(
  zmat,
  re_corr_mat = 1,
  re_sigma,
  ws_corr_mat,
  ws_sigma,
  partition = FALSE,
  print_intermediate = FALSE
)
}
\arguments{
\item{zmat}{Random effects design matrix.}

\item{re_corr_mat}{Random effects correlation matrix (1 if random intercept only).}

\item{re_sigma}{Random effects standard deviation vector.}

\item{ws_corr_mat}{Within-subject correlation matrix, usually identity.}

\item{ws_sigma}{Within-subject standard deviation vector, usually a single scalar.}

\item{partition}{Logical. when \code{TRUE} returns list containing each covariance matrix.}

\item{print_intermediate}{Logical. When \code{TRUE} prints intermediate output.}
}
\value{
A matrix or list of matrices.
}
\description{
Constructs the covariance matrix for linear mixed model based on random effects
and within-subject covariance specifications.
}
\details{
This function builds a subject-specific covariance matrix as
\deqn{\mathbf{\Sigma}_i = \mathbf{Z}_i\mathbf{\Sigma}_b\mathbf{Z}_i^\top + \mathbf{\Sigma}_e}
where \eqn{\mathbf{Z}_i} is the subject-specific random effects matrix, \eqn{\mathbf{\Sigma}_b} is the
random-effects covariance, i.e., constructed from \code{re_corr_mat} and \code{re_sigma},
and \eqn{\mathbf{\Sigma}_e} is the within-subject covariance matrix, which is constructed from
\code{ws_corr_mat} and \code{ws_sigma}. If \code{partition = FALSE}, then only the final matrix
\eqn{\mathbf{Z}_i} is returned, but if \code{partition = TRUE}, then \eqn{\mathbf{Z}_b}
and \eqn{\mathbf{Z}_e} are returned separately in a list.
}
\examples{
xmat_trt <- build_design_matrix(
  N_t = 1,
  N_c = 0,
  K = 5,
  time_start = 0,
  time_scale = 1/4
)

Sigma_i <- build_lmm_cov(
zmat = xmat_trt[, c("intercept", "time")],
re_sigma = c(1.15, 1.05),
re_corr_mat = matrix(
  c(1, 0,
    0, 1),
  byrow = TRUE,
  nrow = 2
),
ws_sigma = 1.25,
ws_corr_mat = diag(5),
print_intermediate = FALSE
)

print(Sigma_i)

Sigma_partition <- build_lmm_cov(
  zmat = xmat_trt[, c("intercept", "time")],
  re_sigma = c(1.15, 1.05),
  re_corr_mat = matrix(
    c(1, 0,
      0, 1),
    byrow = TRUE,
    nrow = 2
  ),
  ws_sigma = 1.25,
  ws_corr_mat = diag(5),
  partition = TRUE,
  print_intermediate = FALSE
)

print(Sigma_partition)

}
