% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcse_estimands.R
\name{mcse_estimands}
\alias{mcse_estimands}
\title{Calculate MCSE Relevant for Estimands}
\usage{
mcse_estimands(
  data,
  true_value = 0,
  estimand_name = "x_diff",
  se_name = "se_diff"
)
}
\arguments{
\item{data}{A data frame where each row contains results from a unique simulated data set.}

\item{true_value}{Numeric. The true value of the estimand.}

\item{estimand_name}{Character. The variable name for the column containing the estimates
of the estimand from each simulated data set.}

\item{se_name}{Character. The variable name for the column containing the estimates
of the SE of the estimator from each simulated data set.}
}
\value{
A data frame.
}
\description{
Combines estimation of MCSE for bias, EmpSE, ModSE, and MSE into a single function.
}
\details{
Combines results from  \link[SimRescueMeds]{mcse_bias}, \link[SimRescueMeds]{mcse_empse},
\link[SimRescueMeds]{mcse_modse},  \link[SimRescueMeds]{mcse_mse} into a single data frame.
}
\examples{
set.seed(283964)
smry_ex <- data.frame(
  beta_hat = rnorm(100),
  se_hat = rgamma(100, shape = 1)
)

mcse_estimands(
  data = smry_ex,
  true_value = 1,
  estimand_name = "beta_hat",
  se_name = "se_hat"
)

mcse_estimands(
  data = smry_ex,
  true_value = 0,
  estimand_name = "beta_hat",
  se_name = "se_hat"
)

}
\references{
\insertRef{Morris2019}{SimRescueMeds}
}
