% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcse_mse.R
\name{mcse_mse}
\alias{mcse_mse}
\title{Calculate MCSE for Mean Squared Error}
\usage{
mcse_mse(measure_obs, true_value, M = NULL, jackknife = FALSE)
}
\arguments{
\item{measure_obs}{A numeric vector where each element is an estimated value from
a simulated data set.}

\item{true_value}{Numeric. The true value of the estimand.}

\item{M}{Numeric. The number of simulated data sets. If \code{NULL}, then calculated
internally.}

\item{jackknife}{Logical. When \code{TRUE}, calculates the jackknife estimate of MCSE based
on Koehler (2009).}
}
\description{
Calculate MCSE for mean squared error of an estimator based on Table 6 of Morris (2019).
}
\details{
This function calculates the Monte Carlo standard error (MCSE) for the mean squared error (MSE)
of an estimator for an estimand from a set of \eqn{M} simulated data sets. If \eqn{\theta} is
the true value of the estimand and \eqn{\hat{\theta}_m} is the estimate in data set \eqn{m},
then estimate of MSE is:
\deqn{
   \widehat{MSE} = \frac{1}{M}\sum_{m = 1}^{M}(\hat{\theta}_m - \theta)^2
}
The MCSE of the MSE is:
\deqn{
   MCSE_{MSE} = \sqrt{\frac{\sum_{m=1}^{M}[(\hat{\theta}_m -\theta)^2 - \widehat{MSE}]^2}{M(M-1)}}
}
}
\examples{
set.seed(283964)
smry_ex <- data.frame(
  beta_hat = rnorm(100),
  se_hat = rgamma(100, shape = 1)
)

mcse_mse(
  measure_obs = smry_ex$beta_hat,
  true_value = 1
)

mcse_mse(
  measure_obs = smry_ex$beta_hat,
  true_value = 1,
  jackknife = TRUE
)

mcse_mse(
  measure_obs = smry_ex$beta_hat,
  true_value = 0
)

mcse_mse(
  measure_obs = smry_ex$beta_hat,
  true_value = 0,
  jackknife = TRUE
)

}
\references{
\insertRef{Koehler2009}{SimRescueMeds}

\insertRef{Morris2019}{SimRescueMeds}
}
