% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcse_rejection.R
\name{mcse_rejection}
\alias{mcse_rejection}
\title{Calculate MCSE for Rejection Rate}
\usage{
mcse_rejection(
  data,
  type = "pvalues",
  alpha = 0.05,
  measure_label = "Rejection"
)
}
\arguments{
\item{data}{A numeric vector of either p-values or binary such that rejected = 1 or not rejected = 0.}

\item{type}{Character. \code{"pvalues"} if \code{data} is a vector of p-values and \code{"rejection"}
if a binary vector.}

\item{alpha}{Numeric. The desired significance level, which should be greater than 0 and less than 1.}

\item{measure_label}{Character. What value should be assigned to column \code{measure}?
Most users will not need to change the default of \code{"Rejection"}.}
}
\description{
Calculate MCSE for rejection rate of a hypothesis test based on Table 6 of Morris (2019).
}
\details{
If If \eqn{\theta} is the true value of the estimand and \eqn{p_m} is a p-value for the hypothesis
test evaluated for simulated data set \eqn{m}, and \eqn{\alpha} is the desired significance level,
then the estimated rejection rate, \eqn{\widehat{Reject}} is the proportion of simulated data sets
in which \eqn{p_m < \alpha}. The MCSE of coverage is then
\deqn{
   MCSE_{Reject} = \sqrt{\frac{\widehat{Reject} \times (1 - \widehat{Reject})}{M}}
}
}
\examples{
set.seed(8972364)
mcse_rejection(
  data = runif(100),
  type = "pvalues"
)

mcse_rejection(
  data = rbinom(100, size = 1, prob = 0.15),
  type = "rejection"
)

}
\references{
\insertRef{Morris2019}{SimRescueMeds}
}
