% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_JMbayes2_rct_long.R
\name{output_JMbayes2_rct_long}
\alias{output_JMbayes2_rct_long}
\title{Extract Longitudinal Summaries for Fitted Bayesian Joint Longitudinal and Time-to-Event Models}
\usage{
output_JMbayes2_rct_long(
  JMbayes2_object,
  JMbayes2_object_summary = FALSE,
  predictor_name = "x:time",
  alpha = 0.05,
  true_beta
)
}
\arguments{
\item{JMbayes2_object}{An An object of class \code{"jm"} or \code{"summary.jm"}, produced by fitting
a joint longitudinal and time-to-event mode \link[JMbayes2]{jm} to analyze a data set.}

\item{JMbayes2_object_summary}{Logical. When \code{TRUE} it is expected that \code{summary()} has alreadly been
applied to the results of fitting \link[JMbayes2]{jm}.}

\item{predictor_name}{Character. The name of the treatment effect as specified in the
\code{formula} argument in \link[nlme]{lme}.}

\item{alpha}{Numeric. The level specified for hypothesis testing and confidence intervals.}

\item{true_beta}{Numeric. The true value of the treatment effect, usually the interaction
between time and treatment assignment.}
}
\description{
Extract treatment effect summaries from the longitudinal sub-model for a single data set
analyzed with joint longitudinal and time-to-event models using \link[JMbayes2]{jm}.
}
\details{
This function is narrowly designed to specifically extract (1) a treatment
effect estimate, usually the interaction between time and treatment group in
longitudinal sub-model, (2) standard errors for the treatment effect,
(3) confidence intervals for the treatment effect, and (4) a p-value for the
treatment effect. Based on the level specified in \code{alpha} argument, whether
the null hypothesis is rejected and whether the CI contains the "true" value are
also included. We additionally output \code{Rhat} for diagnositics.

The idea is to, e.g., use \link[base]{lapply} or \link[purrr]{map}
with this function to apply it to a list of \link[JMbayes2]{jm} objects and produce a
data frame where each row contains the results from a single simulated data set.
}
