% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_lm_rct.R
\name{output_lm_rct}
\alias{output_lm_rct}
\title{Extract Summaries for Fitted Linear Models}
\usage{
output_lm_rct(lm_object, predictor_name = "x", alpha = 0.05, true_beta)
}
\arguments{
\item{lm_object}{An object of class \code{lm}, produce by fitting a linear model using \link[stats]{lm}
to analyze a data set.}

\item{predictor_name}{Character. The name of the treatment effect as specified in the
\code{formula} argument in \link[stats]{lm}.}

\item{alpha}{Numeric. The level specified for hypothesis testing and confidence intervals.}

\item{true_beta}{Numeric. The true value of the treatment effect, usually the interaction
between time and treatment assignment.}
}
\value{
A data frame.
}
\description{
Extract treatment effect summaries for a single data set analyzed with \link[stats]{lm}.
}
\details{
This function is narrowly designed to specifically extract (1) a treatment
effect estimate, (2) standard errors for the treatment effect, (3) confidence
intervals for the treatment effect, and (4) a p-value for the treatment effect.
Based on the level specified in \code{alpha} argument, whether the null hypothesis is
rejected and whether the CI contains the "true" value are also included. The idea is
to, e.g., use \link[base]{lapply} or \link[purrr]{map} with this function to
apply it to a list of \link[stats]{lm} objects and produce a data frame where
each row contains the results from a single simulated data set.
}
