% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_lmm_rct.R
\name{output_lmm_rct}
\alias{output_lmm_rct}
\title{Extract Summaries for Fitted Linear Mixed Models}
\usage{
output_lmm_rct(lme_object, alpha = 0.05, true_beta, predictor_name = "x:time")
}
\arguments{
\item{lme_object}{An object of class \code{lme}, produce by fitting a linear mixed
model using \link[nlme]{lme} to analyze a data set.}

\item{alpha}{Numeric. The level specified for hypothesis testing and confidence intervals.}

\item{true_beta}{Numeric. The true value of the treatment effect, usually the interaction
between time and treatment assignment.}

\item{predictor_name}{Character. The name of the treatment effect as specified in the
\code{formula} argument in \link[nlme]{lme}.}
}
\value{
A data frame.
}
\description{
Extract treatment effect summaries for a single data set analyzed with \link[nlme]{lme}.
}
\details{
This function is narrowly designed to specifically extract (1) a treatment
effect estimate, usually the interaction between time and treatment group,
(2) standard errors for the treatment effect, (3) confidence intervals for
the treatment effect, and (4) a p-value for the treatment effect. Based on
the level specified in \code{alpha} argument, whether the null hypothesis is rejected
and whether the CI contains the "true" value are also included. The idea is
to, e.g., use \link[base]{lapply} or \link[purrr]{map} with this function to
apply it to a list of \link[nlme]{lme} objects and produce a data frame where
each row contains the results from a single simulated data set.
}
\examples{
xmat_t <- build_design_matrix(
    N_t = 1,
    N_c = 0,
    K = 5,
    time_start = 0,
    time_scale = 1/4
    )
xmat_c <- build_design_matrix(
    N_t = 0,
    N_c = 1,
    K = 5,
    time_start = 0,
    time_scale = 1/4
    )

set.seed(2834628)

df <- simulate_lmm_rct(
  N_t = 50,
  N_c = 50,
  xmat_t = xmat_t,
  xmat_c = xmat_c,
  betas = c(
    beta0 = 8,
    beta1 = 0,
    beta2 = -1,
    beta3 = -1
    ),
  Sigma = build_lmm_cov(
    zmat = xmat_t[, c("intercept", "time")],
    re_sigma = c(1.15, 1.05),
    re_corr_mat = matrix(
      c(1, 0,
        0, 1),
      byrow = TRUE,
      nrow = 2
      ),
    ws_sigma = 1.25,
    ws_corr_mat = diag(5),
    print_intermediate = FALSE
    )
)

lme_df <- nlme::lme(
    data = df,
    fixed = y ~ x + time + x:time,
    random = ~ 1 + time | id
)

output_lmm_rct(
  lme_object = lme_df,
  true_beta = -1
)

}
