% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescue_effect_means.R
\name{rescue_effect_means}
\alias{rescue_effect_means}
\title{Produce Mean Rescue Effects as Functions of Time Since Administration}
\usage{
rescue_effect_means(phi0, phi1 = 1, phi, d, decay, delayed_peak = FALSE)
}
\arguments{
\item{phi0}{Numeric. See below for details on interpretation.}

\item{phi1}{Numeric. See below for details on interpretation.}

\item{phi}{Numeric. See below for details on interpretation.}

\item{d}{A numeric vector of times since rescue therapy administration.}

\item{decay}{Character. Specify the functional form of the decay function.
Options are \code{"linear"}, \code{"exponential"}, \code{"gaussian"}.}

\item{delayed_peak}{Logical. When \code{TRUE}, uses functions modified to
allow for a rise, peak, then decay of rescue effect as a function of time
since administration. Must be \code{FALSE} for \code{decay = "linear"}.}
}
\description{
Use (modifications of) linear, exponential, and Gaussian decay functions to
model changing rescue effects as a function of time since rescue therapy
was administered.
}
\details{
This function is designed to produce means for the rescue effect as a function
of time since rescue therapy administration. The underlying idea is add/subtract
a rescue effect from the observed post-rescue therapy longitudinal outcomes to
approximate how rescue therapy affects the observed outcomes. We can draw these
from a distribution, but we would likely expect the rescue effect to be a function
of the time since rescue therapy was administered. The present function produces
mean rescue effects that can then be used to draw randomly from some distribution.
We assume that \eqn{d_t} is the time since rescue therapy was administered and that
\eqn{\delta(d_t)} is mean rescue effect at \eqn{d_t}. In general, we would assume
that rescue effects would decay over time (if not, then let \eqn{\delta(d_t) = \delta}
for all \eqn{d_t}). Linear decay is given by
\deqn{
   \delta_L(d_t) = \phi_0 -\phi d_t
}
Exponential decay is given by
\deqn{
   \delta_E(d_t) = \phi_0\exp\left(-\phi d_t\right)
}
Gaussian decay is given by
\deqn{
   \delta_G(d_t) = \phi_0\exp\left(-\phi d_t^2\right)
}
The parameters \eqn{\phi_0, \phi} are represented by arguments \code{phi0} and \code{phi1},
respectively, and the \code{decay} argument specifies which function to use. If we want to
modify these decay functions to allow for a rise, peak, and decay of rescue effect, then
we modify the exponential and Gaussian decay functions by multiplying them by \eqn{d_t^{\phi_1}}.
\deqn{
\delta_E(d_t) = \phi_0d_t^{\phi_1}\exp(-\phi d_t) \\
\delta_G(d_t) = \phi_0d_t^{\phi_1}\exp(-\phi d_t^2)
}
The argument \code{phi1} specifies the value of \eqn{\phi_1}. Of course, the decay functions above
are reproduced by setting \eqn{\phi_1 = 0}, and setting \eqn{\phi_1 = 1} produces a simplified
setting where \eqn{d_t} is multiplied by the original decay functions. However, we have retained
the argument \code{delayed_peak} so that users who only need decay functions can safely avoid
any foray in these modifications or \code{phi1}.
}
\examples{
d_ex <- seq(0, 1, 0.01)

ld_ex <- rescue_effect_means(phi0 = 1, phi = 2, d = d_ex, decay = "linear")
plot(y = ld_ex, x = d_ex)

ed_ex <- rescue_effect_means(phi0 = 1, phi = 2, d = d_ex, decay = "exponential")
plot(y = ed_ex, x = d_ex)

gd_ex <- rescue_effect_means(phi0 = 1, phi = 2, d = d_ex, decay = "gaussian")
plot(y = gd_ex, x = d_ex)

erpd_ex <- rescue_effect_means(phi0 = 2, phi1 = 1/2, phi = 2, d = d_ex,
                               decay = "exponential", delayed_peak = TRUE)
plot(y = erpd_ex, x = d_ex)

grpd_ex <- rescue_effect_means(phi0 = 2, phi1 = 1/2, phi = 2, d = d_ex,
                               decay = "gaussian", delayed_peak = TRUE)
plot(y = grpd_ex, x = d_ex)

}
