% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_hazard_thomadakis.R
\name{sim_hazard_thomadakis}
\alias{sim_hazard_thomadakis}
\title{Simulate Time-to-Rescue-Therapy for Single Subject}
\usage{
sim_hazard_thomadakis(
  y,
  time,
  theta1 = 0,
  theta2 = 0,
  dist = "exponential",
  lambda = 1,
  gamma = 1,
  maxt = 1
)
}
\arguments{
\item{y}{A vector of longitudinal outcome measurements for a single subject.}

\item{time}{A vector of time points corresponding to each longitudinal measurement.}

\item{theta1}{Numeric. An association parameter for \eqn{y_{i, j}}.}

\item{theta2}{Numeric. An association parameter for \eqn{y_{i,j+1}}. If \code{theta2} is
non-zero, then this implies at least some missingness is not-at-random (MNAR).}

\item{dist}{Character. Either \code{"exponential"} or \code{"weibull"} distributions for the
survival function.}

\item{lambda}{Numeric. The rate parameter for \link[stats]{rexp} or scale for \link[stats]{rweibull}.}

\item{gamma}{Numeric. The shape parameter for \link[stats]{rweibull}.}

\item{maxt}{Numeric. The last possible follow-up time, after which observations
are censored. In the present context, the time the study ends.}
}
\value{
A data frame.
}
\description{
Use proportional hazards models to simulate time-to-rescue therapy
that is dependent on longitudinal trajectories.
}
\details{
This function simulates time-to-rescue medication based on the approach of Thomadakis (2019).
A general formulation for simulating from missing-at-random (MAR) and missing-not-at-random
(MNAR) data is to allow the hazard of rescue therapy within some interval \eqn{t_{i,j} \le t < t_{i,j+1}}
to be a function of the last observed outcome measurement, \eqn{y_{i,j}}, and the next measurement,
\eqn{y_{i,j+1}}, the latter of which is unobserved in the study since it occurs after rescue therapy.
If the event-time distribution is exponential, then we have the following hazard function, but it is
straightforward to extend to Weibull, for which exponential is a special case.

\deqn{
h_i(t) = \lambda\exp\left(y_{i,j}\theta_1 + y_{i,j+1}\theta_2\right), \quad t_{i,j} \le t < t_{i,j+1}.
}

This requires simulating times-to-rescue therapy within each interval between measurements until either
one of those intervals produces rescue therapy or the study ends. If \eqn{\theta_2 \ne 0}, then there is
MNAR missingness. If \eqn{\theta_1 = \theta_2 = 0}, then MCAR. If \eqn{\theta_1 \ne 0} and \eqn{\theta_2 = 0},
then MAR.
}
\examples{
set.seed(65378)
sim_hazard_thomadakis(
  y = rnorm(n = 5),
  time = c(0, 3, 6, 9, 12)/12
  )
sim_hazard_thomadakis(
  y = rnorm(n = 5),
  time = c(0, 3, 6, 9, 12)/12
  )
set.seed(15378)
sim_hazard_thomadakis(
  y = rnorm(n = 5),
  time = c(0, 3, 6, 9, 12)/12
  )

set.seed(28394)
sim_hazard_thomadakis(
  y = rnorm(n = 5),
  time = c(0, 3, 6, 9, 12)/12,
  dist = "weibull",
  gamma = 2
  )

set.seed(28235)
sim_hazard_thomadakis(
  y = rnorm(n = 5),
  time = c(0, 3, 6, 9, 12)/12,
  dist = "weibull",
  gamma = 2
  )

}
\references{
\insertRef{Thomadakis2019}{SimRescueMeds}
}
