% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_hazard_thomadakis_df.R
\name{sim_hazard_thomadakis_df}
\alias{sim_hazard_thomadakis_df}
\title{Simulate Time-to-Rescue-Therapy for a Data Set}
\usage{
sim_hazard_thomadakis_df(
  data,
  id_var = "id",
  y_var = "y",
  time_var = "time",
  theta1 = 0,
  theta2 = 0,
  dist = "exponential",
  lambda = 1,
  gamma = 1,
  maxt = 1,
  print_intermediate = FALSE,
  event_factor = TRUE
)
}
\arguments{
\item{data}{A data frame of longitudinal measurements in long format, i.e.
1 row per observation, not 1 row per subject.}

\item{id_var}{Character. The name of the subject identifier variable.}

\item{y_var}{Character. The name of the longitudinal outcome measurement variable.}

\item{time_var}{Character. The name of the time variable.}

\item{theta1}{Numeric. An association parameter for \eqn{y_{i, j}}.}

\item{theta2}{Numeric. An association parameter for \eqn{y_{i,j+1}}. If \code{theta2} is
non-zero, then this implies at least some missingness is not-at-random (MNAR).}

\item{dist}{Character. Either \code{"exponential"} or \code{"weibull"} distributions for the
survival function.}

\item{lambda}{Numeric. The rate parameter for \link[stats]{rexp} or scale for \link[stats]{rweibull}.}

\item{gamma}{Numeric. The shape parameter for \link[stats]{rweibull}.}

\item{maxt}{Numeric. The last possible follow-up time, after which observations
are censored. In the present context, the time the study ends.}

\item{print_intermediate}{Logical. When \code{TRUE} prints intermediate results.}

\item{event_factor}{Logical. When \code{TRUE} creates a factor variable for the event
indicator, which may be beneficial if no events occur.}
}
\value{
A data frame.
}
\description{
Use proportional hazards models to simulate time-to-rescue therapy
that is dependent on longitudinal trajectories.
}
\details{
This function simulates time-to-rescue medication based on the approach of Thomadakis (2019).
A general formulation for simulating from missing-at-random (MAR) and missing-not-at-random
(MNAR) data is to allow the hazard of rescue therapy within some interval \eqn{t_{i,j} \le t < t_{i,j+1}}
to be a function of the last observed outcome measurement, \eqn{y_{i,j}}, and the next measurement,
\eqn{y_{i,j+1}}, the latter of which is unobserved in the study since it occurs after rescue therapy.
If the event-time distribution is exponential, then we have the following hazard function, but it is
straightforward to extend to Weibull, for which exponential is a special case.

\deqn{
h_i(t) = \lambda\exp\left(y_{i,j}\theta_1 + y_{i,j+1}\theta_2\right), \quad t_{i,j} \le t < t_{i,j+1}.
}

This requires simulating times-to-rescue therapy within each interval between measurements until either
one of those intervals produces rescue therapy or the study ends. If \eqn{\theta_2 \ne 0}, then there is
MNAR missingness. If \eqn{\theta_1 = \theta_2 = 0}, then MCAR. If \eqn{\theta_1 \ne 0} and \eqn{\theta_2 = 0},
then MAR.
}
\examples{
set.seed(3002242)

xmat_t <- build_design_matrix(
    N_t = 1,
    N_c = 0,
    K = 5,
    time_start = 0,
    time_scale = 1/4
  )

df <- simulate_lmm_rct(
  N_t = 50,
  N_c = 50,
  xmat_t = xmat_t,
  xmat_c = build_design_matrix(
    N_t = 0,
    N_c = 1,
    K = 5,
    time_start = 0,
    time_scale = 1/4
  ),
  betas = c(
    beta0 = 8,
    beta1 = 0,
    beta2 = -1,
    beta3 = -1
  ),
  Sigma = build_lmm_cov(
    zmat = xmat_t[, c("intercept", "time")],
    re_sigma = c(1.15, 1.05),
    re_corr_mat = matrix(
      c(1, 0,
        0, 1),
      byrow = TRUE,
      nrow = 2
      ),
    ws_sigma = 1.25,
    ws_corr_mat = diag(5),
    print_intermediate = FALSE
    )
)

set.seed(707582)

sim_hazard_thomadakis_df(
  data = head(df, 15),
  theta1 = 1/3,
  theta2 = 0,
  lambda = 0.05,
  print_intermediate = TRUE
)

sim_hazard_thomadakis_df(
  data = head(df, 15),
  theta1 = 0,
  theta2 = 1/3,
  lambda = 0.05,
  print_intermediate = TRUE,
  dist = "weibull"
)


}
\references{
\insertRef{Thomadakis2019}{SimRescueMeds}
}
