% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_lmm_rct.R
\name{simulate_lmm_rct}
\alias{simulate_lmm_rct}
\title{Simulate Longitudinal RCT Data Set via Linear Mixed Model}
\usage{
simulate_lmm_rct(N_t, N_c, xmat_t, xmat_c, betas, Sigma)
}
\arguments{
\item{N_t}{Numeric. Number of patients that receive active treatment.}

\item{N_c}{Numeric. Number of patients that receive control treatment.}

\item{xmat_t}{A design matrix for those on active treatment.}

\item{xmat_c}{A design matrix for those on control treatment.}

\item{betas}{A numeric vector of fixed effects parameters whose length should equal \code{ncol(xmat)}.}

\item{Sigma}{A covariance matrix. Recommended to generate from \link[SimRescueMeds]{build_lmm_cov}.}
}
\value{
A data frame containing simulated data for a single
data set.
}
\description{
Simulate continuous-valued longitudinal data for RCT using linear mixed model
and subject-specific MVN formulation.
}
\details{
This function simply uses \link[SimRescueMeds]{simulate_subject_MVN} and
\link[base]{replicate} to generate an entire data set. Each treatment group
assumes the same design matrix for all subjects and the dimension of the design
matrices for each group should be equal. That is, they should be followed for
the same number of time points and have the same covariates (usually omitted).
In short, this function is narrowly written for longitudinal RCT-style data.
While subject-specific MVN generation may be slightly more involved to program
than whole-data-set MVN generation, the former is likely to be more computationally
efficient, especially as the sample size increases.
}
\examples{

xmat_t <- build_design_matrix(
    N_t = 1,
    N_c = 0,
    K = 5,
    time_start = 0,
    time_scale = 1/4
    )
xmat_c <- build_design_matrix(
    N_t = 0,
    N_c = 1,
    K = 5,
    time_start = 0,
    time_scale = 1/4
    )

simulate_lmm_rct(
  N_t = 3,
  N_c = 4,
  xmat_t = xmat_t,
  xmat_c = xmat_c,
  betas = c(
    beta0 = 8,
    beta1 = 0,
    beta2 = -1,
    beta3 = -1
    ),
  Sigma = build_lmm_cov(
    zmat = xmat_t[, c("intercept", "time")],
    re_sigma = c(1.15, 1.05),
    re_corr_mat = matrix(
      c(1, 0,
        0, 1),
      byrow = TRUE,
      nrow = 2
      ),
    ws_sigma = 1.25,
    ws_corr_mat = diag(5),
    print_intermediate = FALSE
    )
)


}
