% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_post_tte.R
\name{simulate_post_tte}
\alias{simulate_post_tte}
\title{Modify Post-Rescue Therapy Longitudinal Measurements}
\usage{
simulate_post_tte(
  original_data,
  tte_data,
  id_var = "id",
  y_var = "y",
  time_var = "time",
  tte_var = "eventtime",
  decay_fn = "gaussian"
)
}
\arguments{
\item{original_data}{A data frame containing longitudinal measurements in long format.}

\item{tte_data}{A data frame containing time-to-event measurements with 1 row per subject.}

\item{id_var}{Character. The name of the subject identifier variable. The variable name
should be the same for both \code{original_data} and \code{tte_data} data frames.}

\item{y_var}{Character. The name of the longitudinal outcome measurement variable.}

\item{time_var}{Character. The name of the time variable in the longitudinal data.}

\item{tte_var}{Character. The variable name for the column containing the time-to-event measurements.}
}
\description{
Use a decay function plus noise to modify post-rescue therapy longitudinal
measurements so they are affected postively by rescue medication.
}
\details{
The simulation functions in \code{SimRescueMeds} suggest to begin by simulating
the data that would be observed if no rescue medication occurs. Rescue medication
times are then simulated with (possibly) some dependence on the longitudinal measurements.
However, (believably) simulating "observed" post-rescue therapy longitudinal measurements
requires those measurements to be a function of rescue therapy.
}
