% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_subject_MVN_conditional.R
\name{simulate_subject_MVN_conditional}
\alias{simulate_subject_MVN_conditional}
\title{Simulate Single-Subject Outcomes via Linear Mixed Model}
\usage{
simulate_subject_MVN_conditional(
  xmat,
  betas,
  zmat,
  re_cov_mat,
  ws_cov_mat,
  re_names = NULL,
  print_intermediate = FALSE
)
}
\arguments{
\item{xmat}{A fixed effects design matrix.}

\item{betas}{A numeric vector of fixed effects parameters whose length should equal \code{ncol(xmat)}.}

\item{zmat}{A random effect design matrx.}

\item{re_cov_mat}{A random effects covariance matrix (scalar if only intercept).
This can be obtained from \link[SimRescueMeds]{build_lmm_cov} with \code{partition = TRUE}.}

\item{ws_cov_mat}{A within-subject covariance matrix, usually diagonal. This can be
obtained from \link[SimRescueMeds]{build_lmm_cov} with \code{partition = TRUE}.}

\item{re_names}{An optional vector for variable names for random effects.}

\item{print_intermediate}{Logical. When \code{TRUE} prints intermediate results.}
}
\value{
A data frame.
}
\description{
Draw single-subject data from MVN based on LMM and save random effects.
}
\details{
While \link[MASS]{mvrnorm} is efficient for drawing single-subject outcomes,
it does not explicitly produce or save random effects. For some approaches to
simulating time-to-event data that depends on the longitudinal trajector,
the random effects may be necessary. Here we use \link[MASS]{mvrnorm} to draw
the random effects and within-subject errors separately.
}
\examples{
set.seed(78634)
x_ex <- build_design_matrix(
  N_t = 1,
  N_c = 0,
  K = 5,
  time_start = 0,
  time_scale = 1/4
)

z_ex <- x_ex[, c("intercept", "time")]

cov_ex <- build_lmm_cov(
  zmat = z_ex,
  re_corr_mat = matrix(
    c(1, -0.15,
      -0.15, 1),
    byrow = TRUE, nrow = 2
  ),
  re_sigma = c(2, 3),
  ws_corr_mat = diag(1, 5),
  ws_sigma = 4,
  partition = TRUE,
  print_intermediate = FALSE
)

simulate_subject_MVN_conditional(
  xmat = x_ex,
  zmat = z_ex,
  betas = c(
    beta0 = 8,
    beta1 = 0,
    beta2 = -1,
    beta3 = -1
  ),
  re_cov_mat = cov_ex[["re_cov_mat"]],
  ws_cov_mat = cov_ex[["ws_cov_mat"]],
  print_intermediate = TRUE
)

x_ex0 <- build_design_matrix(
  N_t = 0,
  N_c = 1,
  K = 5,
  time_start = 0,
  time_scale = 1/4
)
z_ex0 <- x_ex[, c("intercept", "time")]

simulate_subject_MVN_conditional(
  xmat = x_ex0,
  zmat = z_ex0,
  betas = c(
    beta0 = 8,
    beta1 = 0,
    beta2 = -1,
    beta3 = -1
  ),
  re_cov_mat = cov_ex[["re_cov_mat"]],
  ws_cov_mat = cov_ex[["ws_cov_mat"]],
  re_names = c("b0", "b2"),
  print_intermediate = TRUE
)

}
